Improve the way automatic decomposition can be used
The user does not have to use the macro @dantzig_wolfe_decompostion(...)
anymore if automatic decomposition is used but decompose(model) can
be called after the constructing all constraints and variables.

Also fix a bug in the plumple method that scores different decompositions.
The "black value" of a block is now computed as v*c where
v is the number of different variables used in the block and
c is the number of constraints in the block