Automatic decomposition (#31)
* axis.jl: compute all possible decompositions

The goal is to implement an automatic decomposition such that the user does not need to provide the decomposition axis.
To find the best axis for the decomposition
1. find the index sets of the constraints (get_constraints_and_axes(...))
2. compute each possible subset of this set of axes; each of these subsets corresponds to one possible decomposition
3. find out how the decomposition/block structure would look like for each of these subsets; for each subset do: get_block_structure(...)
	3.1. built a graph where there is a node for every constraint in the model that is not constructed over a set that is contained in the considered subset
	3.2. find the connected compontents in this graph
	3.3. each connected component will correspond to one block in the decomposition
	3.4. store the possible decomposition as an instance of Block_Structure
To do: 4. implement methods to compare the possible decompositons and determine which block structure should be used for the decomposition

* Enable use of automatic decomposition for Dantzig Wolfe

- BlockDecomposition.jl: allow user to construct a BlockModel where automatic decomposition is activated
- decomposition.jl: modify the function register_decomposition such that an automatic decomposition can be registered
- tree.jl: adapt macro @danzig_wolfe_decomposition for automatic decomposition

* axis.jl: plumple method to score decompositions

Coeffecient matrix should be as white as possible;
master constraints and blocks are black, everything else is white

* Add test file test/automatic_decomposition.jl

* Move automatic decomposition to its own file

It is now in automatic_decomposition.jl and was before in axis.jl.

* Apply style guide blue style

Source: https://github.com/invenia/BlueStyle

* Add test method for automatic decomposition

The test only consists of optimizing the example problem for Coluna
using automatic decomposition.

Note: An error occurs during the execution for an unknown reason.

* Extend tests for automatic decomposition

Includes check for all variables and constraints if they are annotated correctly.

* Annotate variables that occur in a constraint in the same subproblem as the constraint

* Improve the way automatic decomposition can be used

The user does not have to use the macro @dantzig_wolfe_decompostion(...)
anymore if automatic decomposition is used but decompose(model) can
be called after the constructing all constraints and variables.

Also fix a bug in the plumple method that scores different decompositions.
The "black value" of a block is now computed as v*c where
v is the number of different variables used in the block and
c is the number of constraints in the block

* Fix typo in test for automatic decomposition

* Fix bug in src/automatic_decomposition.jl

* Remove return of decomposition and axis when decomposing automatically

The standard value 1 for lower_multiplicity and upper_multiplicity
is used for automatic decomposition. Therefore the user does not
need to access the decomposition tree to set the multiplicities.

* User does not have to call decompose(model) explicitly anymore

Instead, decompose(model::JuMP.Model) is now automatically called in
JuMP.optimize!(model::JuMP.Model). If the user now wants to use
automatic decomposition, all it needs to do is to set
automatic_decomposition to true when creating the block model.

* Rename "plumple" method to "white_score"

* Add the block border score for automatic decomposition

The score is defined in
Khaniyev, Taghi, Samir Elhedhli, and Fatih Safa Erenay.
"Structure detection in mixed-integer programs."
INFORMS Journal on Computing 30.3 (2018): 570-587.

When calling
BlockModel(..., automatic_decomposition=true, automatic_decomposition_score_type=x)
the user can give the type of decomposition score to be used in the
automatic decomposition.
x = 0 -> white score
x = 1 -> block border score from Khaniyev et al.
The standard score is the white score.

* Add the relative border area score for automatic decomposition

The score is defined in
Bergner, Martin, et al. "Automatic Dantzig–Wolfe reformulation of mixed integer programs."
Mathematical Programming 149.1-2 (2015): 391-424.

The user can now choose between 3 different scores:
BlockModel(..., automatic_decomposition=true, automatic_decomposition_score_type=x)
for
x = 0 -> white score
x = 1 -> block border score
x = 2 -> relative border area score

* Restructure the code in automatic_decomposition.jl

* Update Project.toml

Add Metagraphs, LightGraphs, Combinatorics as dependencies.

* Improve structure detection procedure

Now for every subset S of the index sets, two structures
are detected:
1. All constraints index by at least set from S are linking
2. All constraints not indexed by any set from S are linking

Formatting of the code was improved, too.

* Rename functions and files to make clear that Dantzig-Wolfe
is used in automatic decompositon

* Use enum instead of int to choose automatic DW score

To reduce the arguments that need to be given to BlockModel(...)
from two to one (when using automatic DW), the enum type
AutoDwStrategy is introduced.
One can now use the automatic Dantzig-Wolfe decomposition by
creating the model with BlockModel(...,automatic_dantzig_wolfe=x)
where x = white_score | block_border_score | relativ_border_area_score.

* Rename functions to improve readability

* Improve readability of structure detection

Improve readability of the function get_block_structure(...)
which computes for a given set of index sets two candidate structures.

* Support the use of SparseAxisArrays

The automatic decompositon can now deal with
constraints stored in SparseAxisArrays.

* Rename struct holding model information

Rename from Constraints_and_axes to model_description

* Fix bug in automatic decomposition test

Write BlockDecomposition.white_score instead of white_score

* Write JC instead of JuMP.Containers

* Improve comment in automatic Dantzig-Wolfe test

* Fix spelling error

Rename "inaktive" to "inactive"

* Add docstring for scores

* Undo prefixing of names by "BlockDecomposition"

* Remove unnecessary brackets

* Score functions return structure with best score

Instead of the scores of all the structures as
it was the case before. This change was done to improve
readability and structure of the code.

* Change variable name for instance of ModelDescription

* Fix indentation issue

* Add return statements to all functions

* Rename get_block_strucuture to get_block_structures

* Change notation for creation of arrays

* Make ModelDescription struct immutable

Just to make code more elegant: The
description of a model does not change over time.

* Compute connected components in constructor of BlockStructure

* Make annotation of variables more readable

* Use random numbers in the test for the white score

* Implement test for block-border score

* Fix error in relative border area score

Choose the structure with the smallest and not the largest
relative border area score for the decomposition.

* Implement test for relative border area score

Co-authored-by: Guillaume Marques <guillaume.marques@protonmail.com>
Co-authored-by: David Meichel <david.meichel@rwth-aachen.de>